Oguzhan

+++ client/yol-arkadasim/server/controllers/advertController.go
@@ -2,6 +2,7 @@ package controllers
 
 import (
 	"context"
+	"fmt"
 	"net/http"
 	"time"
 	"yol-arkadasim/database"
@@ -14,9 +15,22 @@ import (
 
 // CreateAdvertHandler HTTP POST isteği ile yeni bir ilan oluşturur.
 func CreateAdvertHandler(c *gin.Context) {
-	userID := c.GetString("userID") // Kullanıcı kimliğini middleware'den al
+	var userID any
+	var ok bool
 
-	if c.Request.Method != http.MethodPost {
+	if userID, ok = c.Get("userID"); !ok {
+		c.JSON(http.StatusUnauthorized, gin.H{"error": "User ID not found"})
+		return
+	}
+
+	// userID'yi string türüne dönüştürün
+	userIDStr, ok := userID.(string)
+	if !ok {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Invalid user ID"})
+		return
+	}
+
+	if c.Request.Method != http.MethodPost { // Yöntemi HTTP POST olarak kontrol edin
 		c.JSON(http.StatusMethodNotAllowed, gin.H{"error": "Method Not Allowed"})
 		return
 	}
@@ -30,8 +44,8 @@ func CreateAdvertHandler(c *gin.Context) {
 	// İlan tarihini ayarla
 	now := time.Now()
 	advert.AdvertDate = &now
-	advert.AdvertID = primitive.NewObjectID()                // Yeni ilan ID'si oluştur
-	advert.PostedByID, _ = primitive.ObjectIDFromHex(userID) // Kullanıcı kimliğini ata
+	advert.AdvertID = primitive.NewObjectID() // Yeni ilan ID'si oluştur
+	advert.PostedByID = userIDStr             // Kullanıcı kimliğini ata
 
 	// İlanı veritabanına kaydet
 	if err := SaveAdvertToDB(&advert); err != nil {
@@ -57,10 +71,7 @@ func SaveAdvertToDB(advert *models.AdvertModel) error {
 
 // UpdateAdvertHandler HTTP PUT isteği ile var olan bir ilanı günceller.
 func UpdateAdvertHandler(c *gin.Context) {
-	if c.Request.Method != http.MethodPut {
-		c.JSON(http.StatusMethodNotAllowed, gin.H{"error": "Method Not Allowed"})
-		return
-	}
+	var userID, _ = c.Get("userID")
 
 	// İlan ID'sini URL parametrelerinden al
 	advertIDParam := c.Param("id")
@@ -87,6 +98,17 @@ func UpdateAdvertHandler(c *gin.Context) {
 	client := database.GetMongoClient()
 	collection := client.Database("mydatabase").Collection("adverts")
 
+	var advert models.AdvertModel
+	err = collection.FindOne(context.Background(), bson.M{"advert_id": advertID}).Decode(&advert)
+
+	fmt.Println(userID, advert.PostedByID)
+
+	if userID != advert.PostedByID {
+
+		c.JSON(http.StatusForbidden, gin.H{"error": "bu ilana erişiminiz yok"})
+		return
+	}
+
 	// İlanı güncelle
 	filter := bson.M{"advert_id": advertID}
 	update := bson.M{
@@ -103,7 +125,7 @@ func UpdateAdvertHandler(c *gin.Context) {
 
 	_, err = collection.UpdateOne(context.Background(), filter, update)
 	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Internal Server Error"})
+		c.JSON(http.StatusBadRequest, gin.H{"error": "veritabanına bağlanılamıyor"})
 		return
 	}
 
@@ -143,13 +165,11 @@ func GetAllAdvertsHandler(c *gin.Context) {
 
 // DeleteAdvertHandler HTTP DELETE isteği ile bir ilanı siler.
 func DeleteAdvertHandler(c *gin.Context) {
-	if c.Request.Method != http.MethodDelete {
-		c.JSON(http.StatusMethodNotAllowed, gin.H{"error": "Method Not Allowed"})
-		return
-	}
+
+	var userID, _ = c.Get("userID")
 
 	// İlan ID'sini URL parametrelerinden al
-	advertIDParam := c.Param("id")
+	advertIDParam := c.Param("advert_id")
 	if advertIDParam == "" {
 		c.JSON(http.StatusBadRequest, gin.H{"error": "Advert ID is required"})
 		return
@@ -166,11 +186,21 @@ func DeleteAdvertHandler(c *gin.Context) {
 	client := database.GetMongoClient()
 	collection := client.Database("mydatabase").Collection("adverts")
 
+	var advert models.AdvertModel
+	err = collection.FindOne(context.Background(), bson.M{"advert_id": advertID}).Decode(&advert)
+
+	fmt.Println(advert.PostedByID, userID)
+	if advert.PostedByID != userID {
+		c.JSON(http.StatusForbidden, gin.H{"error": "You can't access this source!"})
+		return
+	}
+
 	// İlanı sil
 	filter := bson.M{"advert_id": advertID}
 	_, err = collection.DeleteOne(context.Background(), filter)
 	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Internal Server Error"})
+		fmt.Println(err.Error())
+		c.JSON(http.StatusBadRequest, gin.H{"error": "Advert silinirken bir hata oluştu."})
 		return
 	}
